latest changes

+++ app/Services/ProcessImageService.php
@@ -6,59 +6,51 @@ use Illuminate\Support\Str;
 
 class ProcessImageService
 {
-    // called function processImage +++++++++++++++++++++++++++++++++++++++
     public function handleProcessImage($imagesString = null, $postId)
     {
-        //dd('In Services: ', $imagesString);
-
-        //  $imageString is a string of image paths seperated by |
-        //  convert the images string parameter into array
-        //  called $explodedImage
         $imagesArray = $imagesString ? explode('|', $imagesString) : [];
-        $userId = Auth()->user('')->id;
-
-        $userId = strval($userId) . '_';
-        $postId = strval($postId) . '_';
-
+        $userId = Auth()->user()->id; // Ensure this is getting the correct user ID
+    
+        $userIdPrefix = strval($userId) . '_';
+        $postIdPrefix = strval($postId) . '_';
+        $identifierPrefix = $userIdPrefix . $postIdPrefix;
+    
         $newImageString = '';
-
-        // dd($userId, $postId, $imagesArray);
-
+    
         if ($imagesArray) {
             foreach ($imagesArray as $image) {
-
-                // Existing files path & file names
-                $storagePathFile = storage_path('app/public/' . $image);
-                //dd($storagePath);
-
-                if (is_file($storagePathFile)) {
-                    // This is for the image copied to the server
-                    // We are adding the user_id of poster and
-                    // the id of the post to track images when
-                    // upload to public server.
-                    $imageSlice = Str::afterLast($image, '/');
-                    $imageOnly = $userId . $postId . $imageSlice;
-
-                    $publicPathFile = public_path(env('POSTS_UPLOAD_PATH') . $imageOnly);
-
-                    //dd($image, $storagePathFile, $publicPathFile);
-
-                    copy($storagePathFile, $publicPathFile);
-                    unlink($storagePathFile);
-                    $newImageString .= $imageOnly . "|";
+    
+                $imageOnly = null;
+                // Extract the filename part after 'uploads/posts/'
+                $imageFilename = Str::after($image, 'uploads/posts/');
+    
+                // Check if the image name already contains the user ID and post ID
+                if (!strpos($image, $identifierPrefix)) {
+                    // Processing required
+                    $storagePathFile = storage_path('app/public/' . $image);
+    
+                    if (is_file($storagePathFile)) {
+                        $imageSlice = Str::afterLast($image, '/');
+                        $imageOnly = $identifierPrefix . $imageSlice;
+                        $publicPathFile = public_path(env('POSTS_UPLOAD_PATH') . $imageOnly);
+    
+                        copy($storagePathFile, $publicPathFile);
+                        unlink($storagePathFile);
+                    }
+                } else {
+                    // No processing required, just use the part of the image name after 'uploads/posts/'
+                    $imageOnly = $imageFilename;
                 }
+                $newImageString .= $imageOnly . "|";
             }
             $newImageString = rtrim($newImageString, '|');
-
         } else {
             $newImageString = null;
         }
-        //dd($newImageString);
-
+    
         return $newImageString;
-
     }
-
+    
     // NOT USING THIS FUNCTION RIGHT NOW
     public function hasImage(string $image): bool
     {
